Reorganises the UI Package
